# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A simple Go API with optional PostgreSQL, Redis, and MongoDB integrations",
        "title": "Simple Go API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/postgres/records": {
            "get": {
                "description": "Get all records from PostgreSQL",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "postgres"
                ],
                "summary": "Get all PostgreSQL records",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/postgres.Record"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new record in PostgreSQL",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "postgres"
                ],
                "summary": "Create PostgreSQL record",
                "parameters": [
                    {
                        "description": "Record data",
                        "name": "record",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/postgres.RecordInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/postgres.Record"
                        }
                    }
                }
            }
        },
        "/postgres/records/{id}": {
            "get": {
                "description": "Get a record from PostgreSQL by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "postgres"
                ],
                "summary": "Get PostgreSQL record by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Record ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/postgres.Record"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a record in PostgreSQL by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "postgres"
                ],
                "summary": "Update PostgreSQL record",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Record ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Record data",
                        "name": "record",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/postgres.RecordInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/postgres.Record"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a record from PostgreSQL by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "postgres"
                ],
                "summary": "Delete PostgreSQL record",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Record ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/redis/cache": {
            "post": {
                "description": "Set a key-value pair in Redis cache",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "redis"
                ],
                "summary": "Set Redis cache item",
                "parameters": [
                    {
                        "description": "Cache item data",
                        "name": "item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/redis.CacheItem"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/redis/cache/{key}": {
            "get": {
                "description": "Get a value from Redis cache by key",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "redis"
                ],
                "summary": "Get Redis cache item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cache key",
                        "name": "key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/redis.CacheItem"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a key from Redis cache",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "redis"
                ],
                "summary": "Delete Redis cache item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cache key",
                        "name": "key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "postgres.Record": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "postgres.RecordInput": {
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "title": {
                    "type": "string"
                }
            }
        },
        "redis.CacheItem": {
            "type": "object",
            "properties": {
                "data": {},
                "expires_in": {
                    "description": "seconds",
                    "type": "integer"
                },
                "key": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        }
    }
}